Split some entry methods into main
This is so they can be unit tested.
